# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api QuickMenu",
        "description": "Documentação da API do sistema de restaurante quickmenu.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/categories": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Store category",
                "description": "Store category",
                "operationId": "d5d39337e0ae1ad328da08f7b468fbbe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Category 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Category 1"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "default": true,
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria registrado com sucesso"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get category",
                "description": "Get category",
                "operationId": "7706bd1bbea7216615ae2d54256aac49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurante encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Category 1"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Category 1"
                                        },
                                        "active": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "restaurant_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoria não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categoria não encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Update category",
                "description": "Update category",
                "operationId": "3876780767a6c0512fcd21fc7913395e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Category 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Category 1"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria atualizado com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete category",
                "description": "Delete category",
                "operationId": "09f09c1bc59c6969c4749771edae84cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria deletada com sucesso"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Category"
                ],
                "summary": "Destroy category",
                "description": "Destroy category",
                "operationId": "6cd45a617d912adf04d79b4b25775c09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria desativada com sucesso"
                    }
                }
            }
        },
        "/api/restaurants/{restaurant}/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get all categories by restaurant",
                "description": "Get all categories by restaurant",
                "operationId": "ad71bf7cc8c0f4d9ff9cfc88779d2e40",
                "parameters": [
                    {
                        "name": "restaurant_id",
                        "in": "path",
                        "description": "ID do restaurante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categorias encontradas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Category 1"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Category 1"
                                            },
                                            "active": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "restaurant_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categorias não encontradas"
                    }
                }
            }
        },
...
    },
...
}